LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/javascript-window-open-blocked-expected.txt

 
FAIL Check that a securitypolicyviolation event is fired assert_equals: expected "inline" but got ""
